requirements.txt
python-jenkins
turtle
boto3
requests
